point to sync to new file
